+++ src/main/java/database/Test.java
@@ -1,4 +1,4 @@
-package parser;
+package database;
 
 import org.hibernate.Session;
 import org.hibernate.Transaction;
